Add initAttestations func to share between nitrite and cose testing
